Better precision in Order: fix transfer for mop
The transfer function for Order (and Marked) used to consider that every
positively marked variable in a point was potentially in the last
position of the buffer. Therefore it flushed every combination of
positively marked x_i after a memory operation on x. Actually, it should
only need to flush the x_i that are alone in their buffer, using the
closed by flushes property of the abstract domain. This passes the
relaxed litmus tests, and should also probably be implemented in
Marked (which might actually carry the same information as the
originally thought Order).