Fixes a mop-flush bug on Abstract
When flushing several variables as specified by a list of tid, the same
initial buffers were considered at each step. Thus the value of the
flushed variable was not properly set to copies who had the same
variable in their buffer in an earlier step (as it was still considered
as present in the buffer).